Add basic p.a.contentrules condition
This condition integrates the IAlert utility within plone.app.contentrules.

This way a contentrule can be used to report about content that
has stop words on it.

The condition allows to either reuse the existing stop words (coming from
the registry) or provide a specific stop words list only for that condition.

It currently only works with comments,
a generic approach will be developed next.