Try to overcome a shortcoming
Problem: there is no relation between a plone.app.contentrules condition
and a string substitution from plone.stringinterp.

This means that if the condition has some data (stop words) that is
important for the string substitution (get stop words snippets) you are
out of luck.

The only away to get around it (as plone.app.discussion is doing) is
to either pass the relevant data around via a session (p.a.discussion <2.3)
 or the request object (p.a.discussion >2.3).