Fix corner case
If the registry returns an empty string or None and no stop_words list
is provided on Alert.get_snippets an empty string needs to be returned.

Not doing so makes the following try to look for an empty string stop word
which unfortunately matches at every single character of the text.