Add a marker interface to objects
Apply it to objects that are reported to have stop words,
and only the ones that are within the scope of a contentrule that
has the ``collective.contentalerts.TextAlert`` condition on it.

This is the building block for a view to show all objects that have
stop words on it.