Split string substitution in two
One for documents and another one for coments.

The problem is that if a comment event triggered a contentrule its
context is the document that contains that comments, thus is impossible
to distinguish between the two of them reliably (i.e between a comment
and its container).

To solve it p.a.discussion should not use the session data manager, which
it does starting from version 2.3.3, but as long as Plone 4.3 is supported
that version can not be used.